Reorganize classes location
