+++ drag-n-drop/app/src/main/java/in/kosacki/dragndropwithnougat/events/NewPathEvent.java
@@ -1,10 +1,11 @@
-package in.kosacki.dragndropwithnougat;
+package in.kosacki.dragndropwithnougat.events;
 
 /**
  * Created by hubert on 27/09/16.
  */
 
 public class NewPathEvent {
+
     private String path;
 
     public NewPathEvent(String pPath){
@@ -14,4 +15,5 @@ public class NewPathEvent {
     public String getPath(){
         return path;
     }
+
 }
